requirements.txt
python-dotenv==0.19.0
hikari[speedups]==2.0.0.dev101
hikari-lightbulb==1.2.1
git+https://github.com/FasterSpeeding/Sake@task/raw-dispatch
apscheduler==3.7.0
aiofiles==0.7.0
asyncpg==0.24.0
uvloop==0.16.0
lavasnek_rs==0.1.0a0